Numbers battle
The program will ask the user to input the size of the array they want to work with and then the program will populate numbers in order starting from zero. The program will then shuffle the numbers, show the odd numbers, even numbers, array number winners, and sorted array winners.