Horse Race
The program will ask the user for the number of horse they want to race. It will then ask for the horse and riders names. The program will then ask the user for the distance they want to see the horses race. The program will ask the user if they would like to see the next second of the race. The race will display > as a horse traversing across the screen until it reaches a | where it will display the winning horse and riders name. The program will then ask the user if they would like to play again.