Counting Delimiters utilizing stacks
The program will prompt the user to enter a sentence. If a left or a right delimiter is used and there is not a matching delimiter in the program, the program will show an error message with a left or right delimiter, character count, and line count and say if it has a missing or mismatching left or right delimiter.